Make sure jiffies value is cast to an integer in the LinuxKPI before doing millisecond conversion. Under FreeBSD jiffies are 32-bit.
MFC after:		1 week
Sponsored by:		Mellanox Technologies


Former-commit-id: 2ad44d0d00d89f481befa90c5439ce17183d1e40

+++ sys/compat/linuxkpi/common/include/linux/jiffies.h
@@ -66,7 +66,7 @@ name|jiffies_to_msecs
 parameter_list|(
 name|x
 parameter_list|)
-value|(((int64_t)(x)) * 1000 / hz)
+value|(((int64_t)(int)(x)) * 1000 / hz)
 end_define
 
 begin_define
